test loadbalancer with ingress and gateway
