+++ tests/integration/tests/conftest.py
@@ -216,7 +216,19 @@ def session_instance(
         bootstrap_config_path,
     )
 
+    instance_default_ip = util.get_default_ip(instance)
+
     instance.exec(["k8s", "bootstrap", "--file", bootstrap_config_path])
+    instance_default_cidr = util.get_default_cidr(instance, instance_default_ip)
+
+    lb_cidr = util.find_suitable_cidr(
+        parent_cidr=instance_default_cidr,
+        excluded_ips=[instance_default_ip],
+    )
+
+    instance.exec(
+        ["k8s", "set", f"load-balancer.cidrs={lb_cidr}", "load-balancer.l2-mode=true"]
+    )
     util.wait_until_k8s_ready(instance, [instance])
     util.wait_for_network(instance)
     util.wait_for_dns(instance)
